requirements.txt
fastapi
uvicorn
pytesseract
python-multipart
pillow
PyMuPDF
